+++ src/oldnews/providers/main.py
@@ -13,6 +13,7 @@ from textual_enhanced.commands import (
 ##############################################################################
 # Local imports.
 from ..commands import (
+    CopyFeedToClipboard,
     CopyHomePageToClipboard,
     Escape,
     MarkAllRead,
@@ -46,6 +47,7 @@ class MainCommands(CommandsProvider):
         yield from self.maybe(OpenHomePage)
         yield from self.maybe(MarkAllRead)
         yield from self.maybe(CopyHomePageToClipboard)
+        yield from self.maybe(CopyFeedToClipboard)
         yield ToggleShowAll()
         yield RefreshFromTheOldReader()
         yield ChangeTheme()